Reorder package files

+++ tests/testthat/test-t12_install_sen2cor.R
@@ -2,6 +2,11 @@ context("Test Sen2Cor installation")
 testthat::skip_on_cran()
 testthat::skip_on_travis()
 
+# NOTE: these tests require a high amount of time (depending on connection speed),
+# so the installation is disabled by default if Sen2Cor is already installed.
+# To perform the download of Sen2Cor, replace 'test_download = FALSE' with 'TRUE'.
+test_download = FALSE
+
 testthat::test_that(
   "Test that a mismatching Sen2Cor version causes error", {
     testthat::expect_error(
@@ -19,10 +24,12 @@ testthat::test_that(
       regexp = "[Oo]nly Sen2Cor versions .+ are currently supported"
     )
     sen2cor_def_version <- package_version("2.8.0")
-    sen2cor_inst_dir <- dirname(dirname(sen2r:::load_binpaths()$sen2cor))
-    unlink(sen2cor_inst_dir, recursive = TRUE)
+    sen2cor_inst_dir <- dirname(dirname(sen2r:::nn(sen2r:::load_binpaths()$sen2cor)))
     
-    install_sen2cor(sen2cor_inst_dir)
+    if (test_download | length(sen2cor_inst_dir) == 0){ 
+      unlink(sen2cor_inst_dir, recursive = TRUE)
+      install_sen2cor(sen2cor_inst_dir)
+    }
     
     testthat::expect_true(dir.exists(dirname(sen2r:::load_binpaths()$sen2cor)))
     testver <- system2(sen2r:::load_binpaths()$sen2cor, "-h", stdout = TRUE)
